Constructionsite display Fixed
bugs in openglcomponent rendering were previously causing errors with the drawing of items at mouse position. the openglcomponent class seemed to be missing one level of scaling for the glviewport and glscissor function. I think I have fixed it but I would like to find the best fix as  I feel there is probably one too many multiplications. For now it works but this is probably the source of future drawing errors?